feat: video profiles
