+++ src/config/profiles/cldAdaptiveStream.json
@@ -17,8 +17,5 @@
         "effect": ["volume:auto"]
       }
     ]
-  },
-  "playlistOptions": {},
-  "playlistByTagOptions": {},
-  "sourcesByTagOptions": {}
+  }
 }
