+++ src/config/profiles/cldDefault.json
@@ -1,0 +1,4 @@
+{
+  "playerOptions": {},
+  "sourceOptions": {}
+}
