+++ src/config/profiles/index.js
@@ -1,0 +1,9 @@
+import cldDefaultProfile from './cldDefault.json';
+import cldLoopingProfile from './cldLooping.json';
+import cldAdaptiveStreamProfile from './cldAdaptiveStream.json';
+
+export const defaultProfiles = {
+  cldDefault: cldDefaultProfile,
+  cldLooping: cldLoopingProfile,
+  cldAdaptiveStream: cldAdaptiveStreamProfile
+};
