+++ src/index.js
@@ -1,6 +1,6 @@
 import 'assets/styles/main.scss';
 import VideoPlayer from './video-player';
-import VideoPlayerProfile from './video-player-profile';
+import createVideoPlayerProfile from './video-player-profile';
 import { assign } from 'utils/assign';
 import { pick, convertKeysToSnakeCase } from './utils/object';
 import { CLOUDINARY_CONFIG_PARAM } from './video-player.const';
@@ -34,9 +34,7 @@ const cloudinaryVideoPlayerConfig = config => ({
 export const videoPlayer = getVideoPlayer();
 export const videoPlayers = getVideoPlayers();
 
-const getVideoPlayerProfile = config => (id, playerOptions, ready) =>
-  new VideoPlayerProfile(id, getConfig(playerOptions, config), ready);
-
+const getVideoPlayerProfile = config => (id, playerOptions, ready) => createVideoPlayerProfile(id, getConfig(playerOptions, config), ready);
 export const videoPlayerProfile = getVideoPlayerProfile();
 
 const cloudinary = {
