--- src/profiles/cldDefault.json
@@ -1,28 +1,0 @@
-{
-  "playerOptions": {
-    "logoOnclickUrl": "https://cloudinary.com/",
-    "showLogo": true,
-    "showJumpControls": false,
-    "skin": "dark",
-    "controls": false,
-    "chaptersButton": false,
-    "pictureInPictureToggle": false,
-    "seekThumbnails": true,
-    "aiHighlightsGraph": false,
-    "preload": "auto",
-    "textTrackSettings": false,
-    "loop": false,
-    "muted": false,
-    "posterOptions": {},
-    "sourceTypes": ["auto"],
-    "floatingWhenNotVisible": "none",
-    "hideContextMenu": false,
-    "analytics": false,
-    "cloudinaryAnalytics": true,
-    "playedEventPercents": [25, 50, 75, 100]
-  },
-  "sourceOptions": {},
-  "playlistOptions": {},
-  "playlistByTagOptions": {},
-  "sourcesByTagOptions": {}
-}
