--- src/utils/fetch-with-timeout.js
@@ -1,14 +1,0 @@
-export async function fetchWithTimeout(resource, options = {}) {
-  const { timeout = 10000 } = options;
-
-  const controller = new AbortController();
-  const id = setTimeout(() => controller.abort(), timeout);
-
-  const response = await fetch(resource, {
-    ...options,
-    signal: controller.signal
-  });
-  clearTimeout(id);
-
-  return response;
-}
